% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_connect}
\alias{scrp_connect}
\title{Interne Funktion: Stellt eine transiente Verbindung mit Retry-Logik her}
\usage{
scrp_connect(sc)
}
\description{
Interne Funktion: Stellt eine transiente Verbindung mit Retry-Logik her
}
\keyword{internal}
