% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_create_table}
\alias{scrp_create_table}
\title{Erstellt eine neue Tabelle in der Datenbank über Spalten und Domains}
\usage{
scrp_create_table(sc, table_name, col_names, domains = "TEXT")
}
\arguments{
\item{sc}{Ein \code{scrp_client}-Objekt.}

\item{table_name}{Name der zu erstellenden Tabelle.}

\item{col_names}{Charakter-Vektor mit den Spaltennamen.}

\item{domains}{Charakter-Vektor mit den Datentypen (z.B. "TEXT", "INTEGER", "AUTO").}
}
\value{
Unsichtbar \code{TRUE} bei Erfolg.
}
\description{
Erstellt eine neue Tabelle in der Datenbank über Spalten und Domains
}
