% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_define_job}
\alias{scrp_define_job}
\title{Erstellt eine Scraping-Job-Konfiguration}
\usage{
scrp_define_job(
  input_table = NULL,
  input_url_column = "url",
  target_tables,
  validate_fn = NULL,
  extract_fn = NULL,
  wait_min_seconds = 5,
  wait_max_seconds = 300,
  batch_size = 1
)
}
\arguments{
\item{input_table}{Optional. Ein Data-Frame oder Tibble. Enthält die
aufzurufenden Ziel-Adressen sowie optionale Metadaten. Falls \code{NULL},
versucht der Executor, die Daten aus der ersten in \code{target_tables}
definierten Tabelle zu laden.}

\item{input_url_column}{Charakter. Der Name der Spalte in der \code{input_table}, welche die
aufzurufenden URLs enthält (Standard: \code{"url"}).}

\item{target_tables}{Eine benannte Liste, die die Ziel-Konfigurationen für die Datenbank enthält.
Jedes Element entspricht einer Tabelle und kann ein Vektor mit Namen der Primärschlüssel-Spalten,
\code{NULL} (für reines Anhängen / Append-Tabellen mit Auto-ID) oder eine Liste mit folgenden Feldern sein:
\describe{
\item{\code{target_key_columns}}{Charakter-Vektor oder \code{NULL}. Die Spaltennamen, die den Primärschlüssel für ein Upsert bilden.}
\item{\code{inherit_input_columns}}{Charakter-Vektor oder \code{NULL}. Namen von Spalten aus der \code{input_table}, die in diese spezifische Zieltabelle übernommen werden sollen (z. B. IDs oder auch die \code{"url"}).}
}}

\item{validate_fn}{Eine optionale Funktion zur Inhalts- und Blockierungsprüfung des HTML-Dokuments.}

\item{extract_fn}{Eine Funktion zur Datenextraktion. Erwartet das HTML-Dokument
der Website (ein \code{xml2::xml_document}) und soll eine benannte Liste von
Tibbles oder Data-Frames zurückgeben. Die Namen der Listenelemente müssen
exakt den Tabellennamen in \code{target_tables} entsprechen.
Die Spaltennamen innerhalb der jeweiligen Data-Frames bilden die Tabellenspalten ab.}

\item{wait_max_seconds}{Numerisch. Maximale Wartezeit in Sekunden (Standard: 300).}

\item{batch_size}{Positive Ganzzahl. Anzahl an URLs, die abgearbeitet werden, bevor
Daten in die Datenbank geschrieben werden (Standard: 1). Bei NULL werden alle
Daten im Arbeitsspeicher gesammelt und erst nach dem Scraping aller URLs in die
Datenbank geschrieben. Dient der Ausbalancierung von Arbeitsspeichernutzung und
Datenbankzugriffen.}
}
\value{
Ein Objekt der Klasse \code{scrp_job}.
}
\description{
\code{scrp_define_job} definiert ein zentrales Konfigurationsobjekt für den Scraper.
Die Funktion steuert, welche URLs verarbeitet werden sollen, wie der Inhalt extrahiert
wird und in welche Datenbanktabellen die Ergebnisse geschrieben werden.
}
