% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_export_csv}
\alias{scrp_export_csv}
\title{Exportiert eine Tabelle aus der Scraper-DB in eine CSV-Datei}
\usage{
scrp_export_csv(sc, table_name, file_path)
}
\arguments{
\item{sc}{Das S3-Scraper-Client-Objekt}

\item{table_name}{Name der Tabelle in der DB}

\item{file_path}{Pfad zur Ziel-CSV}
}
\description{
Exportiert eine Tabelle aus der Scraper-DB in eine CSV-Datei
}
