% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_has_redirect}
\alias{scrp_has_redirect}
\title{Check, ob Weiterleitung stattgefunden hat beim Aufruf einer Website}
\usage{
scrp_has_redirect(expected_url, actual_url)
}
\arguments{
\item{expected_url}{Die URL der angesteuerten Website}

\item{actual_url}{Die tatsächliche URL nach Aufruf der Website}
}
\description{
Check, ob Weiterleitung stattgefunden hat beim Aufruf einer Website
}
