% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_run_job}
\alias{scrp_run_job}
\title{High-Level Pipeline: Führt einen Scraping-Job aus}
\usage{
scrp_run_job(sc, job)
}
\arguments{
\item{sc}{Ein \code{scrp_client}-Objekt.}

\item{job}{Ein \code{scrp_job}-Objekt, erstellt durch \code{scrp_define_job}.}
}
\value{
Unsichtbar \code{TRUE} bei Erfolg.
}
\description{
High-Level Pipeline: Führt einen Scraping-Job aus
}
