% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_setup_browser}
\alias{scrp_setup_browser}
\title{Richtet den Pfad für den simulierten Browser (chromote) ein}
\usage{
scrp_setup_browser(path = NULL)
}
\arguments{
\item{path}{Optionaler, spezifischer Pfad zur ausführbaren Datei des Browsers (z.B. chrome.exe oder msedge.exe).
Falls NULL, sucht die Funktion nach Standardpfaden für Chrome, Edge und Brave.}
}
\value{
Der gesetzte Pfad (unsichtbar).
}
\description{
Richtet den Pfad für den simulierten Browser (chromote) ein
}
