% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_client}
\alias{scrp_client}
\title{Scraper Client erstellen (S3-Klasse): Ermöglicht Datenbankzugriff und Logging}
\usage{
scrp_client(con, log_table_name = "log", use_fake_browser = FALSE)
}
\arguments{
\item{con}{Ein bestehendes DBI-Verbindungsobjekt (z.B. zu SQLite)}

\item{log_table_name}{Name der Logging-Tabelle in der DB}
}
\value{
Ein S3-Objekt mit Datenbankverbindung und Kennung der Logging-Tabelle
}
\description{
Scraper Client erstellen (S3-Klasse): Ermöglicht Datenbankzugriff und Logging
}
