% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscrape.R
\name{scrp_log_status}
\alias{scrp_log_status}
\title{API-Schnittstelle: Status/Log in die DB schreiben}
\usage{
scrp_log_status(sc, url, status, http_status = NA_character_)
}
\arguments{
\item{sc}{Das S3-Scraper-Client-Objekt}

\item{url}{Die URL der betreffenden Seite}

\item{status}{Der sinngemäße Status des Scraping-Versuchs (z.B. "success", "not found", "unavailable", "blocked", "redirected", "missing data")}

\item{http_status}{Der Status der http request}
}
\description{
API-Schnittstelle: Status/Log in die DB schreiben
}
